main-domain/backend/requirements.txt
# Core Flask dependencies
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0
Flask-Bcrypt>=1.0.0
Flask-Limiter>=3.0.0
Flask-Mail>=0.9.1

# Database
PyMySQL>=1.0.0
SQLAlchemy>=2.0.0

# Security
python-dotenv>=1.0.0
cryptography>=3.4.8
Werkzeug>=2.3.0

# Validation
marshmallow>=3.19.0
email-validator>=2.0.0
phonenumbers>=8.12.0

# Development and testing
pytest>=7.0.0
pytest-flask>=1.2.0
coverage>=7.0.0

# Production
gunicorn>=21.0.0
redis>=4.5.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
requests>=2.32.0

# Image processing
Pillow>=9.0.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2
